修改内容

+++ Shoprize_EN/Shoprize_EN/VIMapViewController.h
@@ -19,7 +19,7 @@
     
     MKMapView *map;
     
-    int currentIndex;
+    long currentIndex;
     
 }
